# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorAction": {
        "enum": [
          "retry",
          "reload",
          "wait",
          "view_diff",
          "open_resync",
          "unlock_vault",
          "open_provider_settings",
          "edit_instruction",
          "change_model",
          "adjust_budget"
        ],
        "title": "ErrorAction",
        "type": "string"
      },
      "ErrorCode": {
        "enum": [
          "VALIDATION",
          "REVISION_CONFLICT",
          "WORK_BLOCKED",
          "WORK_BUSY_QUEUED",
          "WORK_ACTIVE_JOB",
          "CHAPTER_RANGE_CONFLICT",
          "CHAPTER_READ_ONLY",
          "CANDIDATE_NOT_READY",
          "CANDIDATE_CLOSED",
          "CHAPTER_IS_BASE",
          "STATE_SNAPSHOT_MISSING",
          "CHAPTER_EMPTY",
          "VAULT_LOCKED",
          "VAULT_PASSWORD_INVALID",
          "VAULT_UNLOCK_THROTTLED",
          "VAULT_ALREADY_EXISTS",
          "VAULT_NOT_FOUND",
          "VAULT_CORRUPT",
          "SECRET_MISSING",
          "PROVIDER_AUTH",
          "PROVIDER_UNREACHABLE",
          "PROVIDER_SERVER_ERROR",
          "PROVIDER_RATE_LIMIT",
          "PROVIDER_IN_USE",
          "PROVIDER_REFUSAL",
          "OUTPUT_TRUNCATED",
          "STRUCTURED_OUTPUT_INVALID",
          "BUDGET_EXCEEDED",
          "UNAUTHORIZED",
          "FORBIDDEN_HOST",
          "FORBIDDEN_ORIGIN",
          "NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "DB_BUSY",
          "BACKEND_SHUTTING_DOWN",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorAction"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LanguageFindingOut": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message_key": {
            "title": "Message Key",
            "type": "string"
          },
          "needs_confirmation": {
            "title": "Needs Confirmation",
            "type": "boolean"
          },
          "paragraph_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph Id"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion"
          }
        },
        "required": [
          "check_id",
          "kind",
          "severity",
          "confidence",
          "quote",
          "message_key",
          "params",
          "needs_confirmation"
        ],
        "title": "LanguageFindingOut",
        "type": "object"
      },
...
      "NormalizeIn": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "enum": [
              "paste",
              "save"
            ],
            "title": "Mode",
            "type": "string"
          },
          "text": {
            "maxLength": 1000000,
            "title": "Text",
            "type": "string"
          },
          "work_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Id"
          }
        },
        "required": [
          "text",
          "mode"
        ],
        "title": "NormalizeIn",
        "type": "object"
      },
      "NormalizeOut": {
        "properties": {
          "changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Changes",
            "type": "array"
          },
          "legacy_encoding": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy Encoding"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "changes"
        ],
        "title": "NormalizeOut",
        "type": "object"
      },
...
      "ParagraphTextIn": {
        "additionalProperties": false,
        "properties": {
          "paragraph_id": {
            "pattern": "^[a-z0-9]{8}$",
            "title": "Paragraph Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "paragraph_id",
          "text"
        ],
        "title": "ParagraphTextIn",
        "type": "object"
      },
...
      "SlopEntryIn": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "is_regex": {
            "default": false,
            "title": "Is Regex",
            "type": "boolean"
          },
          "max_per_1000_units": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Per 1000 Units"
          },
          "note": {
            "default": "",
            "maxLength": 500,
            "title": "Note",
            "type": "string"
          },
          "pattern": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Pattern",
            "type": "string"
          },
          "scope": {
            "default": "anywhere",
            "enum": [
              "paragraph_start",
              "anywhere"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pattern"
        ],
        "title": "SlopEntryIn",
        "type": "object"
      },
      "SlopListOut": {
        "properties": {
          "builtin": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Builtin",
            "type": "array"
          },
          "user": {
            "additionalProperties": true,
            "title": "User",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "builtin",
          "user",
          "version"
        ],
        "title": "SlopListOut",
        "type": "object"
      },
      "SlopListPut": {
        "additionalProperties": false,
        "properties": {
          "additions": {
            "items": {
              "$ref": "#/components/schemas/SlopEntryIn"
            },
            "maxItems": 500,
            "title": "Additions",
            "type": "array"
          },
          "disabled": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "title": "Disabled",
            "type": "array"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "SlopListPut",
        "type": "object"
      },
...
      "TextCheckIn": {
        "additionalProperties": false,
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "checks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks"
          },
          "paragraphs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParagraphTextIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraphs"
          }
        },
        "title": "TextCheckIn",
        "type": "object"
      },
      "TextCheckOut": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/LanguageFindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "length": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Length",
            "type": "object"
          }
        },
        "required": [
          "findings",
          "length"
        ],
        "title": "TextCheckOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WriteStoryApp backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/languages/{code}/genre-presets": {
      "get": {
        "operationId": "language_genre_presets",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Language Genre Presets",
        "tags": [
          "languages"
        ]
      }
    },
...
    "/v1/languages/{code}/normalize": {
      "post": {
        "operationId": "normalize_language_text",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NormalizeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Normalize Language Text",
        "tags": [
          "languages"
        ]
      }
    },
    "/v1/languages/{code}/slop-list": {
      "get": {
        "operationId": "get_slop_list",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlopListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Slop List",
        "tags": [
          "languages"
        ]
      },
      "put": {
        "operationId": "put_slop_list",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlopListPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Put Slop List",
        "tags": [
          "languages"
        ]
      }
    },
...
    "/v1/works/{work_id}/text/check": {
      "post": {
        "operationId": "check_work_text",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextCheckOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Check Work Text",
        "tags": [
          "languages"
        ]
      }
    }
  }
}